requirements.txt
# Core dependencies
fastapi
uvicorn

# Embedding and vector search
langchain==0.0.208
openai==0.27.4
pinecone-client==2.2.2

# Configuration and environment management
pydantic
python-dotenv==1.0.0

# Logging
loguru==0.6.0

# Optional for testing and linting
pytest==7.3.1
black==23.3.0
